+++ app/api/protokoll/route.ts
@@ -28,7 +28,7 @@ const postSchema = z.object({
     .transform((v) => (v === "" || v === undefined ? null : v)),
   anfahrt_km: z.preprocess(
     (v) => (v === "" || v === undefined ? undefined : v === null ? null : v),
-    z.union([z.coerce.number().int().min(0), z.null()]).optional()
+    z.union([z.coerce.number().min(0), z.null()]).optional()
   ),
   anfahrt_minuten: z.preprocess(
     (v) => (v === "" || v === undefined ? undefined : v === null ? null : v),
